requirements.txt
affine==2.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.2
blis==1.3.3
bs4==0.0.2
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cligj==0.7.2
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contextily==1.7.0
contourpy==1.3.3
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.13
de_dep_news_trf @ https://github.com/explosion/spacy-models/releases/download/de_dep_news_trf-3.8.0/de_dep_news_trf-3.8.0-py3-none-any.whl#sha256=6da411d41d5b0465beb1725549039e05caf00fec9ae1dc41794c8f7555824905
distro==1.9.0
filelock==3.20.1
fonttools==4.61.0
fsspec==2025.12.0
geographiclib==2.1
geopandas==1.1.2
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mercantile==1.2.1
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
numpy==2.3.5
openai==2.14.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
preshed==3.0.12
pydantic==2.12.5
pydantic_core==2.41.5
pyogrio==0.12.1
pyparsing==3.2.5
pyproj==3.7.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
rasterio==1.5.0
regex==2025.11.3
requests==2.32.5
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sympy==1.14.0
thinc==8.3.10
torch==2.9.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.3
wheel==0.45.1
wrapt==2.0.1
xyzservices==2025.11.0